improve installextpkg

+++ installextpkg.sh
@@ -9,18 +9,14 @@ tdir=~/Tools
 ddir="${tdir}/download"
 idir="${tdir}/install"
 
-######################################################################
-# 設定
-######################################################################
-
 print_usage_and_exit () {
   cat <<-USAGE 1>&2
 	Usage   : ${0##*/}
-	Options : -c
+	Options : -f
 
 	必要な環境をインストールする。
 
-	-cオプションでシステムにコマンドをインストールする（要管理者権限）。
+	-fオプションで既存のディレクトリを削除して再インストールできる。
 	USAGE
   exit 1
 }
@@ -30,14 +26,14 @@ print_usage_and_exit () {
 ######################################################################
 
 opr=''
-opt_c='no'
+opt_f='no'
 
 i=1
 for arg in ${1+"$@"}
 do
   case "$arg" in
     -h|--help|--version) print_usage_and_exit ;;
-    -c)                  opt_c='yes'          ;;
+    -f)                  opt_f='yes'          ;;
     *)
       if [ $i -eq $# ] && [ -z "$opr" ]; then
         opr=$arg
@@ -51,7 +47,7 @@ do
   i=$((i + 1))
 done
 
-iscmdinstall=$opt_c
+isreinstall=$opt_f
 
 ######################################################################
 # 事前準備
@@ -60,22 +56,31 @@ iscmdinstall=$opt_c
 mkdir -p "$ddir"
 mkdir -p "$idir"
 
+#####################################################################
+# 前準備
+######################################################################
+
+if [ "$isreinstall" == 'yes' ]; then
+  # 既存のツールを削除
+  cd "$idir"
+
+  [ -d 'shellshoccar' ] && rm -rf 'shellshoccar'
+fi
+
 ######################################################################
 # 本体処理
 ######################################################################
 
 # shellshoccar
 (
-  # 既存のツールを削除
-  cd "$idir"
-  [ -d 'shellshoccar' ] && rm -rf 'shellshoccar'
-
-  # ツールを配置
   cd "$ddir"
-  [ -d 'shellshoccar' ] && rm -rf 'shellshoccar'
-  git clone https://github.com/ShellShoccar-jpn/installer.git shellshoccar
-  cd shellshoccar
-  sh shellshoccar.sh --prefix="${idir}/shellshoccar" install
+
+  # インストール本体処理
+  if [ ! -d 'shellshoccar' ]; then
+    git clone https://github.com/ShellShoccar-jpn/installer.git shellshoccar
+    cd shellshoccar
+    sh shellshoccar.sh --prefix="${idir}/shellshoccar" install
+  fi
 
   # パスを追加
   estr='export PATH="'"${idir}/shellshoccar/bin:"'${PATH}''"'
@@ -83,10 +88,3 @@ mkdir -p "$idir"
     echo "$estr" >> ~/.bashrc
   fi
 )
-
-# コマンドインストール
-if [ "$iscmdinstall" = 'yes' ]; then
-  sudo apt install -y vim screen
-  sudo apt install -y gawk
-  sudo apt install -y nkf
-fi
